+++ cmd/create/create-topics.go
@@ -19,6 +19,6 @@ var cmdCreateTopic = &cobra.Command{
 func init() {
 	cmdCreateTopic.Flags().Int32VarP(&flags.Partitions, "partitions", "p", 1, "number of partitions")
 	cmdCreateTopic.Flags().Int16VarP(&flags.ReplicationFactor, "replication-factor", "r", 1, "replication factor")
-	cmdCreateTopic.Flags().BoolVarP(&flags.ValidateOnly, "validate-only", "V", false, "validate only")
-	cmdCreateTopic.Flags().StringArrayVarP(&flags.Configs, "config", "C", flags.Configs, "configs in format `key=value`")
+	cmdCreateTopic.Flags().BoolVarP(&flags.ValidateOnly, "validate-only", "v", false, "validate only")
+	cmdCreateTopic.Flags().StringArrayVarP(&flags.Configs, "config", "c", flags.Configs, "configs in format `key=value`")
 }